New Rendering API & Colour type change to 0-1 range

+++ SC/src/Renderer/OpenGL/OpenGLFrameBuffer.cpp
@@ -59,7 +59,7 @@ namespace SC
 		GLCall(glGenTextures(1, &tex));
 		GLCall(glBindTexture(GL_TEXTURE_2D, tex));
 
-		GLCall(glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB8, w, h, 0, GL_RGB, GL_UNSIGNED_BYTE, nullptr));
+		GLCall(glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, w, h, 0, GL_RGB, GL_UNSIGNED_BYTE, nullptr));
 
 		GLCall(glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR));
 		GLCall(glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR));
